% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puntomedio.R, R/rectangulo.R, R/simpson.R,
%   R/trapecio.R
\name{f}
\alias{f}
\title{Integración por la regla el punto medio}
\usage{
f(x)

f(x)

f(x)

f(x)
}
\arguments{
\item{f}{funcion a integrar, hay que meterla aparte}

\item{a}{un punto}

\item{b}{otro punto}
}
\value{
la integral de la funcion entre a y b

la integral de la funcion entre a y b

la integral de la funcion entre a y b

la integral de la funcion entre los puntos a y b
}
\description{
Integración por la regla el punto medio

Integración por la regla del rectangulo

Integracion por la regla de simpson

Integracion con la regla del trapecio
}
\examples{
pmedio(x^3,0,2)
rectangulo(x^2,0,1)
simpson(f,0,1)
trapecio(x^3,1,3)
}
