% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puntomedio.R
\name{pmedio}
\alias{pmedio}
\title{Integración por la regla el punto medio}
\usage{
pmedio(f, a, b)
}
\arguments{
\item{f}{una funcion}

\item{a}{un punto}

\item{b}{otro punto}
}
\value{
la integral de la funcion entre a y b
}
\description{
Integración por la regla el punto medio
}
\examples{
pmedio(x^3,0,2)
}
