% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangulo.R
\name{rectangulo}
\alias{rectangulo}
\title{Integración por la regla del rectangulo}
\usage{
rectangulo(f, a, b)
}
\arguments{
\item{f}{una función que queramos integrar}

\item{a}{un punto}

\item{b}{otro punto}
}
\value{
la integral de la funcion entre a y b
}
\description{
Integración por la regla del rectangulo
}
\examples{
rectangulo(x^2,0,1)
}
