% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regla_rectangulo.R
\name{regla_rectangulo}
\alias{regla_rectangulo}
\title{Integración mediante la Regla del Rectángulo}
\usage{
regla_rectangulo(a, b, f)
}
\arguments{
\item{a}{extremo inferior del intervalo}

\item{b}{extremo superior del intervalo}

\item{f}{función que vamos a integrar}
}
\value{
La aproximación de la integral en el intervalo
}
\description{
Esta función nos calcula una aproximación de la integral de f en el intervalo
de a hasta b, mediante la Regla del Rectángulo.
}
\details{
En nuestra entrada f, tenemos que definir nuestra función. Es decir
f es del tipo función.
}
\examples{
regla_rectangulo(1.5,3,function(x){return(x^2)})
regla_rectangulo(2,3,function(x){return(sin(x))})

}
\seealso{
\link{regla_punto_medio}
}
