% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regla_simpson.R
\name{regla_simpson}
\alias{regla_simpson}
\title{Integración mediante la Regla de Simpson}
\usage{
regla_simpson(a, b, f)
}
\arguments{
\item{a}{extremo inferior del intervalo}

\item{b}{extremo superior del intervalo}

\item{f}{función que vamos a integrar}
}
\value{
La aproximación de la integral en el intervalo
}
\description{
Esta función nos calcula una aproximación de la integral de f en el intervalo
de a hasta b, mediante la Regla de Simpson.
}
\details{
En nuestra entrada f, tenemos que definir nuestra función. Es decir
f es del tipo función.
}
\examples{
regla_simpson(1,3,function(x){return(x^2+x^3+x)})
regla_simpson(2,4,function(x){return(cos(x)+x)})

}
\seealso{
\link{regla_punto_medio}, \link{regla_rectangulo}, \link{regla_trapecio}
}
