% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson.R
\name{simpson}
\alias{simpson}
\title{Integracion por la regla de simpson}
\usage{
simpson(f, a, b)
}
\arguments{
\item{f}{funcion a integrar}

\item{a}{un punto}

\item{b}{otro punto}
}
\value{
la integral de la funcion entre a y b
}
\description{
Integracion por la regla de simpson
}
\examples{
simpson(x^2,0,1)
}
