% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapecio.R
\name{trapecio}
\alias{trapecio}
\title{Integracion con la regla del trapecio}
\usage{
trapecio(f, a, b)
}
\arguments{
\item{f}{funcion a integrar}

\item{a}{un punto}

\item{b}{otro punto}
}
\value{
la integral de la funcion entre los puntos a y b
}
\description{
Integracion con la regla del trapecio
}
\examples{
trapecio(x^3,1,3)
}
